requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
stable-baselines3==2.3.0
gymnasium==0.29.1
yfinance==0.2.40
pandas==2.2.0
numpy==1.26.0
pandas-ta==0.3.14b
scikit-learn==1.4.0
python-dotenv==1.0.0
